Function graph_graphql::execution::execution::execute_selection_set

fn execute_selection_set<'a>(
    ctx: &'a ExecutionContext<impl Resolver>, 
    selection_sets: impl Iterator<Item = &'a SelectionSet>, 
    object_type: &ObjectType, 
    prefetched_value: Option<Value>
) -> Result<Value, Vec<QueryExecutionError>>

Executes a selection set, requiring the result to be of the given object type.

Allows passing in a parent value during recursive processing of objects and their fields.