Module graph_graphql::prelude

Prelude that exports the most important traits and types.

Re-exports

pub use super::query::execute_query;

pub use super::query::ext::BlockConstraint;

pub use super::query::QueryExecutionOptions;

pub use super::schema::api_schema;

pub use super::schema::ast::validate_entity;

pub use super::schema::APISchemaError;

pub use super::subscription::SubscriptionExecutionOptions;

Structs

ExecutionContext
Contextual information passed around during query execution.

GraphQlRunner
GraphQL runner implementation for The Graph.

IntrospectionResolver

Query
A GraphQL query that has been preprocessed and checked and is ready for execution. Checking includes validating all query fields and, if desired, checking the query’s complexity

StoreResolver
A resolver that fetches entities from a Store.

Traits

MaybeCoercible
A GraphQL value that can be coerced according to a type.

Resolver
A GraphQL resolver that can resolve entities, enum values, scalar types and interfaces/unions.

Functions

build_query
Builds a EntityQuery from GraphQL arguments.

introspection_schema

Type Definitions

ObjectType