Trait graph_graphql::prelude::Resolver

pub trait Resolver: Sized + Send + Sync + 'static {
    const CACHEABLE: bool;

#[must_use]    fn query_permit<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = OwnedSemaphorePermit> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn prefetch(
        &self, 
        ctx: &ExecutionContext<Self>, 
        selection_set: &SelectionSet
    ) -> Result<Option<Value>, Vec<QueryExecutionError>>;

    fn resolve_objects(
        &self, 
        prefetched_objects: Option<Value>, 
        field: &Field, 
        field_definition: &Field, 
        object_type: ObjectOrInterface<'_>, 
        arguments: &HashMap<&str, Value>
    ) -> Result<Value, QueryExecutionError>;

    fn resolve_object(
        &self, 
        prefetched_object: Option<Value>, 
        field: &Field, 
        field_definition: &Field, 
        object_type: ObjectOrInterface<'_>, 
        arguments: &HashMap<&str, Value>
    ) -> Result<Value, QueryExecutionError>;

    fn resolve_enum_value(
        &self, 
        _field: &Field, 
        _enum_type: &EnumType, 
        value: Option<Value>
    ) -> Result<Value, QueryExecutionError> { ... }

    fn resolve_scalar_value(
        &self, 
        _parent_object_type: &ObjectType, 
        _field: &Field, 
        _scalar_type: &ScalarType, 
        value: Option<Value>, 
        _argument_values: &HashMap<&str, Value>
    ) -> Result<Value, QueryExecutionError> { ... }

    fn resolve_enum_values(
        &self, 
        _field: &Field, 
        _enum_type: &EnumType, 
        value: Option<Value>
    ) -> Result<Value, Vec<QueryExecutionError>> { ... }

    fn resolve_scalar_values(
        &self, 
        _field: &Field, 
        _scalar_type: &ScalarType, 
        value: Option<Value>
    ) -> Result<Value, Vec<QueryExecutionError>> { ... }

    fn resolve_abstract_type<'a>(
        &self, 
        schema: &'a Document, 
        _abstract_type: &TypeDefinition, 
        object_value: &Value
    ) -> Option<&'a ObjectType> { ... }

#[must_use]    fn resolve_field_stream<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        _schema: &'life1 Document, 
        _object_type: &'life2 ObjectType, 
        _field: &'life3 Field
    ) -> Pin<Box<dyn Future<Output = Result<StoreEventStreamBox, QueryExecutionError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn post_process(&self, _result: &mut QueryResult) -> Result<(), Error> { ... }
}

A GraphQL resolver that can resolve entities, enum values, scalar types and interfaces/unions.

Associated Constants

const CACHEABLE: bool

Loading content...
Required methods

#[must_use]fn query_permit<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = OwnedSemaphorePermit> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn prefetch(
&self,
ctx: &ExecutionContext<Self>,
selection_set: &SelectionSet
) -> Result<Option<Value>, Vec<QueryExecutionError>>

Prepare for executing a query by prefetching as much data as possible

fn resolve_objects(
&self,
prefetched_objects: Option<Value>,
field: &Field,
field_definition: &Field,
object_type: ObjectOrInterface<'_>,
arguments: &HashMap<&str, Value>
) -> Result<Value, QueryExecutionError>

Resolves list of objects, prefetched_objects is Some if the parent already calculated the value.

fn resolve_object(
&self,
prefetched_object: Option<Value>,
field: &Field,
field_definition: &Field,
object_type: ObjectOrInterface<'_>,
arguments: &HashMap<&str, Value>
) -> Result<Value, QueryExecutionError>

Resolves an object, prefetched_object is Some if the parent already calculated the value.

Loading content...
Provided methods

fn resolve_enum_value(
&self,
_field: &Field,
_enum_type: &EnumType,
value: Option<Value>
) -> Result<Value, QueryExecutionError>

Resolves an enum value for a given enum type.

fn resolve_scalar_value(
&self,
_parent_object_type: &ObjectType,
_field: &Field,
_scalar_type: &ScalarType,
value: Option<Value>,
_argument_values: &HashMap<&str, Value>
) -> Result<Value, QueryExecutionError>

Resolves a scalar value for a given scalar type.

fn resolve_enum_values(
&self,
_field: &Field,
_enum_type: &EnumType,
value: Option<Value>
) -> Result<Value, Vec<QueryExecutionError>>

Resolves a list of enum values for a given enum type.

fn resolve_scalar_values(
&self,
_field: &Field,
_scalar_type: &ScalarType,
value: Option<Value>
) -> Result<Value, Vec<QueryExecutionError>>

Resolves a list of scalar values for a given list type.

fn resolve_abstract_type<'a>(
&self,
schema: &'a Document,
_abstract_type: &TypeDefinition,
object_value: &Value
) -> Option<&'a ObjectType>

#[must_use]fn resolve_field_stream<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
_schema: &'life1 Document,
_object_type: &'life2 ObjectType,
_field: &'life3 Field
) -> Pin<Box<dyn Future<Output = Result<StoreEventStreamBox, QueryExecutionError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

fn post_process(&self, _result: &mut QueryResult) -> Result<(), Error>

Loading content...