Module graph_graphql::query::ast

Utilities for working with GraphQL query ASTs.

Functions

get_argument_value
Looks up the value of an argument in a vector of (name, value) tuples.

get_directive
Looks up a directive in a selection, if it is provided.

get_fragment
Returns up the fragment with the given name, if it exists.

get_operation
Returns the operation for the given name (or the only operation if no name is defined).

get_operation_name
Returns the name of the given operation (if it has one).

get_operations
Returns all operation definitions in the document.

get_response_key
Returns the response key of a field, which is either its name or its alias (if there is one).

get_variable_definitions
Returns the variable definitions for an operation.

include_selection
Returns true if a selection should be included (as per the @include directive).

skip_selection
Returns true if a selection should be skipped (as per the @skip directive).