Trait graph_graphql::query::ext::ValueExt

pub trait ValueExt: Sized {
    fn as_object(&self) -> &BTreeMap<String, Value>;

    fn as_string(&self) -> &str;

    fn lookup<'a>(
        &'a self, 
        vars: &'a HashMap<String, Self>, 
        pos: Pos
    ) -> Result<&'a Self, QueryExecutionError>;
}

Required methods

fn as_object(&self) -> &BTreeMap<String, Value>

fn as_string(&self) -> &str

fn lookup<'a>(
&'a self,
vars: &'a HashMap<String, Self>,
pos: Pos
) -> Result<&'a Self, QueryExecutionError>

If self is a variable reference, look it up in vars and return that. Otherwise, just return self.

If self is a variable reference, but has no entry in vars return an error

Loading content...