Function graph_graphql::store::prefetch::fetch

fn fetch(
    logger: Logger, 
    store: &impl QueryStore + ?Sized, 
    parents: &Vec<&mut Node>, 
    join: &Join<'_>, 
    arguments: HashMap<&str, Value>, 
    multiplicity: ChildMultiplicity, 
    types_for_interface: &BTreeMap<EntityType, Vec<ObjectType>>, 
    block: BlockNumber, 
    max_first: u32, 
    max_skip: u32, 
    query_id: String, 
    collected_column_names: BTreeMap<ObjectCondition<'_>, AttributeNames>
) -> Result<Vec<Node>, QueryExecutionError>

Query child entities for parents from the store. The join indicates in which child field to look for the parent’s id/join field. When is_single is true, there is at most one child per parent.