Function graph_graphql::store::prefetch::run

pub fn run(
    resolver: &StoreResolver, 
    ctx: &ExecutionContext<impl Resolver>, 
    selection_set: &SelectionSet
) -> Result<Value, Vec<QueryExecutionError>>

Run the query in ctx in such a manner that we only perform one query per ‘level’ in the query. A query like musicians { id bands { id } } will perform two queries: one for musicians, and one for bands, regardless of how many musicians there are.

The returned value contains a q::Value::Object that contains a tree of all the entities (converted into objects) in the form in which they need to be returned. Nested object fields appear under the key r:response_key in these objects, and are always q::Value::List of objects.

For the above example, the returned object would have one entry under r:musicians, which is a list of all the musicians; each musician has an entry r:bands that contains a list of the bands for that musician. Note that even for single-object fields, we return a list so that we can spot cases where the store contains data that violates the data model by having multiple values for what should be a relationship to a single object in @derivedFrom fields